requirements.txt
requests==2.32.2
flightradarapi==1.3.34
psycopg2-binary==2.9.10
dash==3.0.2
plotly==6.0.1
pandas==2.2.2
python-dateutil==2.9.0post0
schedule==1.2.1
